+++ packages/terra-date-picker/src/DatePickerField.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Field from 'terra-form-field';
+import { injectIntl, intlShape } from 'react-intl';
 import IconError from 'terra-icon/lib/icon/IconError';
 import DatePicker from './DatePicker';
 
@@ -47,6 +48,11 @@ const propTypes = {
    * Do not set the name in inputAttribute as it will be ignored.
    */
   inputAttributes: PropTypes.object,
+  /**
+   * @private
+   * intl object programmatically imported through injectIntl from react-intl.
+   * */
+  intl: intlShape.isRequired,
   /**
   * Whether the field displays as Incomplete. Use when no value has been provided. _(usage note: `required` must also be set)_.
   */
@@ -188,6 +194,7 @@ const DatePickerField = (props) => {
     isInline,
     isLabelHidden,
     includeDates,
+    intl,
     label,
     labelAttrs,
     maxDate,
@@ -227,13 +234,27 @@ const DatePickerField = (props) => {
     mergedInputAttrs = { 'aria-describedby': ariaDescriptionIds, ...inputAttributes };
   }
 
+  const format = intl.formatMessage({ id: 'Terra.datePicker.dateFormat' });
+
+  const helpLabel = help ? (
+    <div id="format" aria-label={`${intl.formatMessage({ id: 'Terra.datePicker.dateFormatLabel' })} ${format}, ${help}`}>
+      {`(${format})`}
+      &nbsp;
+      {help}
+    </div>
+  ) : (
+    <div id="format" aria-label={`${intl.formatMessage({ id: 'Terra.datePicker.dateFormatLabel' })} ${format}`}>
+      {`(${format})`}
+    </div>
+  );
+
   return (
     <Field
       {...customProps}
       label={label}
       labelAttrs={labelAttrs}
       error={error}
-      help={help}
+      help={helpLabel}
       hideRequired={hideRequired}
       required={required}
       showOptional={showOptional}
@@ -249,6 +270,7 @@ const DatePickerField = (props) => {
         inputAttribute={mergedInputAttrs}
         excludeDates={excludeDates}
         filterDate={filterDate}
+        useExternalFormatMask
         includeDates={includeDates}
         isInvalid={isInvalid}
         isIncomplete={isIncomplete}
@@ -274,4 +296,4 @@ const DatePickerField = (props) => {
 DatePickerField.propTypes = propTypes;
 DatePickerField.defaultProps = defaultProps;
 
-export default DatePickerField;
+export default injectIntl(DatePickerField);