+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -4,6 +4,15 @@
 
 * Changed
   * `onClickOutside` prop has been deprecated, will resolve to `onRequestClose`
+  * Updated wdio snapshots due to changes in date-picker and time-input
+
+* Changed
+  * Update `timeVariant` prop description to include `es` in locales supporting only 24 hour format.
+
+## 4.46.0 - (September 10, 2020)
+
+* Fixed
+  * Fix to clear out invalid date value when time input selected in date-time-picker.
 
 ## 4.45.0 - (August 18, 2020)